add permission functionality

+++ backend/src/resolvers/Mutation.js
@@ -3,14 +3,24 @@ const jwt = require('jsonwebtoken');
 const { transport, makeANiceEmail } = require('../mail');
 const { randomBytes} = require('crypto');
 const { promisify } = require('util');
+const { hasPermission } = require('../utils')
 
 
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!')
+    }
     const item = await ctx.db.mutation.createItem(
       {
         data: {
+          // this is how we create relationship between the Item and the User
+          user:  {
+            connect: {
+              id: ctx.request.userId
+            }
+          },
           ...args
         }
       },
@@ -36,11 +46,17 @@ const Mutations = {
     );
   },
   async deleteItem(parent, args, ctx, info) {
+    // throw new Error('You arrent allowed!!!!')
     const where = { id: args.id };
     // find the item
-    const item = await ctx.db.query.item({ where }, `{id, title}`);
+    const item = await ctx.db.query.item({ where }, `{id, title user { id }}`);
     // Check if they own the item or have the permissions
-    //TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission))
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error('You don\'t have permission to do that')
+    }
     // delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -97,25 +113,7 @@ const Mutations = {
     ctx.response.clearCookie('token');
     return { message: 'Goodbye!'};
   },
-  //  async requestReset( parent,args, ctx, info ) {
-  //   //1. check if real user
-  //   const user = await ctx.db.query.user({ where : { email: args.email }});
 
-  //   if(!user) {
-  //     throw new Error(`No such user found for email ${args.email}`)
-  //   }
-  //   //2. set a reset token and expiry on that user
-  //   const randomBytesPromiseified = promisify(randomBytes);
-  //   const resetToken = (await randomBytesPromiseified(20)).toString('hex');
-  //   const resetTokenExpiry = Data.now() + 3600000; // 1 houre from now
-  //   const res = await ctx.db.mutation.updateUser({
-  //     where: { email: args.email},
-  //     data: { resetToken, resetTokenExpiry}
-  //   });
-  //   console.log(res);
-  //   return { message: 'Thanks!'};
-  //   //3. Email them that user token 
-  // }
   async requestReset(parent, args, ctx, info) {
     // 1. Check if this is a real user
     const user = await ctx.db.query.user({ where: { email: args.email } });
@@ -183,6 +181,34 @@ const Mutations = {
     return updatedUser
     // 9. have a beer
 
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. check if they are logged in
+    if(!ctx.request.userId) {
+      throw new Error('You must be logged in!')
+      }
+    // 2. Query the qurrent user
+     const currentUser = await ctx.db.query.user({
+       where: {
+         id: ctx.request.userId,
+       }
+     }, info
+     );
+    // 3. Check if they have permmissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE'])
+    // 4. Update the permissions
+    return ctx.db.mutation.updateUser({
+      data: {
+        permissions: {
+          set: args.permissions,
+        }
+      },
+      where: {
+        id: args.userId,
+      },
+
+    }, info)
   }
 };
